[EFL] Change data type of parameter in ewk_view_viewport_attributes_get(). https://bugs.webkit.org/show_bug.cgi?id=69505
Reviewed by Andreas Kling.

Source/WebKit/efl:

ewk_view_viewport_attributes_get() has used implicit casting for width and height of viewport.
However, it is better to use *int* data type as parameter data type instead of *float* casting.

* ewk/ewk_view.cpp:
(ewk_view_viewport_attributes_get):
* ewk/ewk_view.h:

Tools:

* EWebLauncher/main.c: Use *int* variables for width and height of viewport meta tag.
(on_viewport_changed):


Canonical link: https://commits.webkit.org/85648@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96890 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/efl/ewk/ewk_view.cpp
@@ -3456,7 +3456,7 @@ void ewk_view_viewport_attributes_set(Evas_Object* o, const WebCore::ViewportArg
     evas_object_smart_callback_call(o, "viewport,changed", 0);
 }
 
-void ewk_view_viewport_attributes_get(const Evas_Object* o, float* w, float* h, float* init_scale, float* max_scale, float* min_scale, float* device_pixel_ratio, Eina_Bool* user_scalable)
+void ewk_view_viewport_attributes_get(const Evas_Object* o, int* w, int* h, float* init_scale, float* max_scale, float* min_scale, float* device_pixel_ratio, Eina_Bool* user_scalable)
 {
     WebCore::ViewportAttributes attributes = _ewk_view_viewport_attributes_compute(o);
